move color palette to separate file
