+++ src/utils/Theme/themes/colorPalette.ts
@@ -1,0 +1,168 @@
+
+export const colorPalette = {
+  white: '#FFFFFF',
+  black: '#000000',
+  gray: {
+    100: '#F7F7FA',
+    200: '#E6E6F0',
+    300: '#C9C9D6',
+    400: '#77778C',
+    500: '#5C5C73',
+    600: '#3D3D4D',
+    700: '#242433',
+    800: '#1E1E2E',
+    900: '#11111F',
+    950: '#02020F'
+  },
+  magenta: {
+    100: '#FAE5F6',
+    200: '#F1BFE9',
+    300: '#E480D3',
+    400: '#D740BE',
+    500: '#C900A8',
+    600: '#AB018F',
+    700: '#8C0276',
+    800: '#6E025C',
+    900: '#4F0343',
+    950: '#31042A'
+  },
+  pink: {
+    100: '#FDE5F1',
+    200: '#F9BFDB',
+    300: '#F480B7',
+    400: '#EE4094',
+    500: '#DE006B',
+    600: '#BB015A',
+    700: '#98014A',
+    800: '#740239',
+    900: '#510229',
+    950: '#2E0318'
+  },
+  lime: {
+    100: '#F4FAE5',
+    200: '#E3F3BF',
+    300: '#C6E880',
+    400: '#AADC40',
+    500: '#8ED000',
+    600: '#78B001',
+    700: '#628F01',
+    800: '#4C6F02',
+    900: '#364E02',
+    950: '#202E03'
+  },
+  teal: {
+    100: '#E5FBF9',
+    200: '#BFF6F0',
+    300: '#80EDE0',
+    400: '#40E5D1',
+    500: '#00DCC2',
+    600: '#00C2AB',
+    700: '#019A88',
+    800: '#017365',
+    900: '#024B42',
+    950: '#02231F'
+  },
+  cyan: {
+    100: '#E5F9FE',
+    200: '#BFF0FB',
+    300: '#80E2F8',
+    400: '#40D3F4',
+    500: '#00C5F0',
+    600: '#01A7CB',
+    700: '#0289A6',
+    800: '#036B82',
+    900: '#044D5D',
+    950: '#052F38'
+  },
+  violet: {
+    100: '#F0E8FD',
+    200: '#DAC5F9',
+    300: '#B58BF3',
+    400: '#9152EE',
+    500: '#6C18E8',
+    600: '#5B14C5',
+    700: '#4B10A1',
+    800: '#3A0D7E',
+    900: '#2A095B',
+    950: '#190537'
+  },
+  purple: {
+    100: '#F3E5FC',
+    200: '#E2BFF7',
+    300: '#C580F0',
+    400: '#A840E8',
+    500: '#8B00E0',
+    600: '#7501BC',
+    700: '#5F0298',
+    800: '#490274',
+    900: '#330350',
+    950: '#1D042D'
+  },
+  red: {
+    100: '#FCE5E6',
+    200: '#F7BFC1',
+    300: '#F08083',
+    400: '#E84045',
+    500: '#E00007',
+    600: '#B70006',
+    700: '#8E0005',
+    800: '#660104',
+    900: '#3D0103',
+    950: '#200204'
+  },
+  orange: {
+    100: '#FEF3E5',
+    200: '#FDE1BF',
+    300: '#FBC280',
+    400: '#F8A340',
+    500: '#F68500',
+    600: '#CB6E00',
+    700: '#9F5701',
+    800: '#743F01',
+    900: '#482802',
+    950: '#251602'
+  },
+  yellow: {
+    100: '#FFF9E5',
+    200: '#FFF1BF',
+    300: '#FFE380',
+    400: '#FFD440',
+    500: '#FFC600',
+    600: '#D2A300',
+    700: '#A58001',
+    800: '#785E01',
+    900: '#4B3B02',
+    950: '#261F03'
+  },
+  green: {
+    100: '#EEF8E9',
+    200: '#D5EFC8',
+    300: '#AADF91',
+    400: '#80CE5B',
+    500: '#55BE24',
+    600: '#469D1D',
+    700: '#377C16',
+    800: '#275C10',
+    900: '#183B09',
+    950: '#091A02'
+  },
+  blue: {
+    100: '#E7EFFF',
+    200: '#C3D7FF',
+    300: '#87AEFF',
+    400: '#4C86FF',
+    500: '#105EFF',
+    600: '#0D4ED2',
+    700: '#0A3DA6',
+    800: '#082D79',
+    900: '#051C4C',
+    950: '#041028'
+  },
+  'black-pearl': '#02020F',
+  'athens-gray': '#F7F7FA',
+  mystic: '#E6E6F0',
+  vulcan: '#11111F',
+  charade: '#242433',
+  waterloo: '#77778C',
+  anakiwa: '#93B6FF'
+};
