+++ src/utils/Theme/themes/index.ts
@@ -2,3 +2,4 @@ export * from './theme';
 export * from './extendTheme';
 export * from './extendCoponentTheme';
 export * from './extractTheme';
+export * from './colorPalette';
